# perf: Add `ArrayChunks` to optimize codegen of BatchDecoder