--- a/crates/polars-parquet/src/arrow/read/deserialize/primitive/mod.rs
+++ b/crates/polars-parquet/src/arrow/read/deserialize/primitive/mod.rs
@@ -3,7 +3,9 @@ mod dictionary;
 mod integer;
 mod nested;
 
-pub use basic::Iter;
+pub(crate) use basic::{
+    AsDecoderFunction, DecoderFunction, IntoDecoderFunction, Iter, UnitDecoderFunction,
+};
 pub use dictionary::{DictIter, NestedDictIter};
 pub use integer::IntegerIter;
 pub use nested::NestedIter;