--- a/crates/polars-parquet/src/arrow/read/deserialize/utils/mod.rs
+++ b/crates/polars-parquet/src/arrow/read/deserialize/utils/mod.rs
@@ -1,5 +1,6 @@
 use std::collections::VecDeque;
 
+pub(crate) mod array_chunks;
 pub(crate) mod filter;
 
 use arrow::array::{BinaryArray, MutableBinaryViewArray, View};