… crates/polars-parquet/src/arrow/read/deserialize/primitive/dictionary.rs
… crates/polars-parquet/src/arrow/read/deserialize/primitive/integer.rs
… crates/polars-parquet/src/arrow/read/deserialize/primitive/nested.rs
… crates/polars-parquet/src/arrow/read/deserialize/simple.rs
… crates/polars-parquet/src/arrow/read/deserialize/utils/array_chunks.rs
… crates/polars-parquet/src/parquet/types.rs