This increases visibility and allows for better control of inlining.

From the first commit we get the following benchmark.

```
Benchmark 1: After Optimization
  Time (mean ± σ):     18.095 s ±  0.135 s    [User: 10.459 s, System: 7.564 s]
  Range (min … max):   17.976 s … 18.412 s    10 runs

Benchmark 2: Before Optimization
  Time (mean ± σ):     20.863 s ±  0.041 s    [User: 13.129 s, System: 7.648 s]
  Range (min … max):   20.794 s … 20.918 s    10 runs

Summary
  After Optimization ran
    1.15 ± 0.01 times faster than Before Optimization
```